// pslpwcl/src/app/modules/pslpapi/model/parametroResponse.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApiMessage } from './apiMessage';
import { Parametro } from './parametro';

export interface ParametroResponse { 
    apiMessages?: Array<ApiMessage>;
    esitoPositivo?: boolean;
    parametro?: Parametro;
}